# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SwipeJobs API - Tinder/TikTok for Jobs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: Get job feed for swiping (auth optional — guests get all active jobs, authenticated users get personalized feed)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          description: Full-text search across job title, description, and tags
          schema:
            type: string
        - name: jobType
          in: query
          schema:
            type: string
            enum: [fulltime, contract, parttime, internship]
        - name: experienceLevel
          in: query
          schema:
            type: string
            enum: [junior, mid, senior, lead, executive, any]
        - name: remoteType
          in: query
          schema:
            type: string
            enum: [remote, hybrid, onsite]
      responses:
        "200":
          description: Job feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details (auth optional)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Admin - list all users
      security:
        - clerkAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
  /admin/jobs:
    get:
      operationId: adminGetJobs
      tags: [admin]
      summary: Admin - list all jobs
      security:
        - clerkAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
    post:
      operationId: adminCreateJob
      tags: [admin]
      summary: Admin - create a job
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /admin/jobs/{jobId}:
    patch:
      operationId: adminUpdateJob
      tags: [admin]
      summary: Admin - update a job
      security:
        - clerkAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: adminDeleteJob
      tags: [admin]
      summary: Admin - delete a job
      security:
        - clerkAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/analytics:
    get:
      operationId: adminGetAnalytics
      tags: [admin]
      summary: Admin - get platform analytics
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analytics"
  /admin/jobs/import:
    post:
      operationId: adminImportJobs
      tags: [admin]
      summary: Admin - trigger job import
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
  /admin/ats/career-urls:
    get:
      operationId: adminGetCareerUrls
      tags: [admin]
      summary: Admin - list company career URLs for ATS import
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Career URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerUrlsResponse"
    post:
      operationId: adminAddCareerUrl
      tags: [admin]
      summary: Admin - add a company career URL
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCareerUrlRequest"
      responses:
        "201":
          description: Career URL added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerUrl"
  /admin/ats/import/{urlId}:
    post:
      operationId: adminImportFromAts
      tags: [admin]
      summary: Admin - import jobs from a specific ATS career URL
      security:
        - clerkAuth: []
      parameters:
        - name: urlId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResponse"
  /admin/ats/career-urls/{urlId}:
    delete:
      operationId: adminDeleteCareerUrl
      tags: [admin]
      summary: Admin - remove a career URL
      security:
        - clerkAuth: []
      parameters:
        - name: urlId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
        role:
          type: string
          enum: [job_seeker, employer, admin]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive]
          nullable: true
        preferredLocation:
          type: string
          nullable: true
        remotePreference:
          type: string
          enum: [remote, hybrid, onsite, any]
          nullable: true
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        jobCategories:
          type: array
          items:
            type: string
        onboardingCompleted:
          type: boolean
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - role
        - onboardingCompleted
        - isAdmin
        - jobCategories
        - createdAt
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required:
        - id
        - name
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          $ref: "#/components/schemas/Company"
        location:
          type: string
          nullable: true
        remoteType:
          type: string
          enum: [remote, hybrid, onsite]
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        jobType:
          type: string
          enum: [fulltime, parttime, contract, internship]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive, any]
        tags:
          type: array
          items:
            type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        applyUrl:
          type: string
        source:
          type: string
        matchScore:
          type: integer
          nullable: true
          description: Match percentage 0-100 based on user preferences
        isSaved:
          type: boolean
        isPaidListing:
          type: boolean
        viewCount:
          type: integer
        createdAt:
          type: string
        expiresAt:
          type: string
          nullable: true
      required:
        - id
        - title
        - company
        - remoteType
        - jobType
        - experienceLevel
        - tags
        - shortDescription
        - fullDescription
        - applyUrl
        - source
        - isSaved
        - isPaidListing
        - viewCount
        - createdAt
...
    JobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        hasMore:
          type: boolean
      required:
        - jobs
        - total
        - hasMore
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
      required:
        - users
        - total
    CreateJobRequest:
      type: object
      properties:
        title:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        location:
          type: string
        remoteType:
          type: string
          enum: [remote, hybrid, onsite]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        jobType:
          type: string
          enum: [fulltime, parttime, contract, internship]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive, any]
        tags:
          type: array
          items:
            type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        applyUrl:
          type: string
      required:
        - title
        - remoteType
        - jobType
        - experienceLevel
        - shortDescription
        - fullDescription
        - applyUrl
    UpdateJobRequest:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        remoteType:
          type: string
          enum: [remote, hybrid, onsite]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        jobType:
          type: string
          enum: [fulltime, parttime, contract, internship]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive, any]
        tags:
          type: array
          items:
            type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        applyUrl:
          type: string
        isActive:
          type: boolean
    Analytics:
      type: object
      properties:
        totalUsers:
          type: integer
        totalJobs:
          type: integer
        totalSwipes:
          type: integer
        totalSaved:
          type: integer
        totalClicks:
          type: integer
        totalEmployers:
          type: integer
        swipesLast7Days:
          type: integer
        newUsersLast7Days:
          type: integer
        clicksLast7Days:
          type: integer
        jobsImportedLast7Days:
          type: integer
      required:
        - totalUsers
        - totalJobs
        - totalSwipes
        - totalSaved
        - totalClicks
        - totalEmployers
        - swipesLast7Days
        - newUsersLast7Days
        - clicksLast7Days
        - jobsImportedLast7Days
    ImportResponse:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        message:
          type: string
      required:
        - imported
        - message
    CareerUrl:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        careerUrl:
          type: string
        atsType:
          type: string
          enum: [greenhouse, lever, workable, generic]
        lastImportedAt:
          type: string
          nullable: true
        lastImportCount:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - companyName
        - careerUrl
        - atsType
        - createdAt
    CareerUrlsResponse:
      type: object
      properties:
        careerUrls:
          type: array
          items:
            $ref: "#/components/schemas/CareerUrl"
        total:
          type: integer
      required:
        - careerUrls
        - total
    AddCareerUrlRequest:
      type: object
      properties:
        companyName:
          type: string
        careerUrl:
          type: string
        atsType:
          type: string
          enum: [greenhouse, lever, workable, generic]
      required:
        - companyName
        - careerUrl
        - atsType