# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwipeJobs API - Tinder/TikTok for Jobs
 * OpenAPI spec version: 0.1.0
 */

export interface Analytics {
  totalUsers: number;
  totalJobs: number;
  totalSwipes: number;
  totalSaved: number;
  totalClicks: number;
  totalEmployers: number;
  swipesLast7Days: number;
  newUsersLast7Days: number;
  clicksLast7Days: number;
  jobsImportedLast7Days: number;
}